rename QuantumTape -> JacobianTape and BaseQuantumTape -> QuantumTape
