+++ pennylane/tape/tapes/__init__.py
@@ -15,8 +15,8 @@
 This subpackage contains various quantum tapes, which track, queue,
 validate, execute, and differentiate quantum circuits.
 """
-from .base import BaseQuantumTape
-from .jacobian_tape import QuantumTape
+from .tape import QuantumTape
+from .jacobian_tape import JacobianTape
 from .cv_param_shift import CVParamShiftTape
 from .qubit_param_shift import QubitParamShiftTape
 from .reversible import ReversibleTape
