+++ pennylane/tape/tapes/reversible.py
@@ -23,13 +23,13 @@ import numpy as np
 
 import pennylane as qml
 
-from .jacobian_tape import QuantumTape
+from .jacobian_tape import JacobianTape
 
 
 ABC_ARRAY = np.array(list(ABC))
 
 
-class ReversibleTape(QuantumTape):
+class ReversibleTape(JacobianTape):
     r"""Quantum tape for computing gradients via reversible analytic differentiation.
 
     .. note::
@@ -48,7 +48,7 @@ class ReversibleTape(QuantumTape):
 
     >>> tape.jacobian(dev)
 
-    For more details on the quantum tape, please see :class:`~.QuantumTape`.
+    For more details on the quantum tape, please see :class:`~.JacobianTape`.
 
     **Reversible analytic differentiation**
 
@@ -181,7 +181,7 @@ class ReversibleTape(QuantumTape):
 
         generator = generator(wires)
 
-        diff_circuit = QuantumTape()
+        diff_circuit = JacobianTape()
         diff_circuit._ops = [copy(op).inv() for op in between_ops[::-1]] + [generator] + between_ops
 
         # set the simulator state to be the pre-measurement state
